02/backend/requirements.txt
Flask==2.3.2
Pillow==10.0.0
flask-cors==4.0.0
gunicorn==21.2.0